curand now returns Float32, qr_fix is allowscalar compliant

+++ src/cuda/wigner.jl
@@ -1,4 +1,4 @@
 function curand(w::WignerEnsemble{β}) where β
     z = curand(w.g)
-    (z + z') / 2sqrt(2β * w.d)
+    (z + z') / 2sqrt(2β * Float32(w.d))
 end
